Refactor search service
